+++ include/agent47.hpp
@@ -8,16 +8,11 @@
 #ifdef AGENT47_HAS_ROS2
     #include "agent47/bridge/ros2_bridge.hpp"
 #endif
-#include "agent47/model/entity.hpp"
-#include "agent47/model/robot.hpp"
 #include "agent47/sensor.hpp"
 #include "agent47/sensors/gnss.hpp"
 #include "agent47/sensors/imu.hpp"
 #include "agent47/sensors/lidar.hpp"
 #include "agent47/types.hpp"
+#include <datapod/robot.hpp>
 
 #include "agent47/model/urdf.hpp"
-
-#ifdef AGENT47_HAS_ROS2
-    #include "agent47/bridge/ros2_bridge.hpp"
-#endif